refactor(rs-api): extract delivery_monitor to free file-size budget (#174)
Review finding 🟡#5. delivery.rs was at exactly 1000 lines (the CI cap)
after fix #4; any addition for Task 14 / 15 would re-trip the gate.
Move monitor_delivery_health (~150 lines) into a sibling module via
`impl DeliveryOrchestrator` block. Same behavior; uses pool() and
audit_tx() accessors instead of private field reads. DELIVERY_RL now
pub(crate) so the sibling module can reach it.

delivery.rs goes 999 -> 846 lines; ~150 lines of headroom.
